refactor(support-chat): extrair SupportInboxView e useSupportInboxState compartilhados
Remove duplicação entre SupportInbox (admin) e PartnerSupportInbox (cursinho).
Fix: aside com max-w-96 evita que scrollbar cubra o UnreadBadge dos itens.
